atomic design created with components
